combined 2 menus into one, add style
